+++ scripts/lichessTools-init.js
@@ -31,6 +31,7 @@ lichessTools.loadTool(LiChessTools.Tools.Themes);
 lichessTools.loadTool(LiChessTools.Tools.CliCommands);
 lichessTools.loadTool(LiChessTools.Tools.ChessOps);
 lichessTools.loadTool(LiChessTools.Tools.D3);
+lichessTools.loadTool(LiChessTools.Tools.NoUiSlider);
 lichessTools.loadTool(LiChessTools.Tools.ExportPGN);
 lichessTools.loadTool(LiChessTools.Tools.Stockfish);
 lichessTools.loadTool(LiChessTools.Tools.AdditionalGlyphs);